Trying to fix logging...
